Add receiving current user's location and get data about current weather with user's location

+++ lib/data/api/model/current_weather/api_coord.dart
@@ -1,6 +1,6 @@
 class ApiCoord {
-  int? lon;
-  int? lat;
+  double? lon;
+  double? lat;
 
   ApiCoord({this.lon, this.lat});
